Update App and Profile tests to use fakeStore
