+++ test/components/Profile.test.js
@@ -1,6 +1,8 @@
 const expect = require('chai').expect;
 const React = require('react');
+import { Provider } from 'react-redux';
 import { mount } from 'enzyme';
+import fakeStore from './fakeStore';
 
 import Profile from '../../client/components/Profile.js';
 
@@ -10,18 +12,24 @@ const fakeProfile = {
 };
 
 describe('<Profile />', () => {
+  let wrapper;
+
+  beforeEach('render Profile', () => {
+    wrapper = mount(
+      <Provider store={fakeStore}>
+        <Profile />
+      </Provider>);
+  });
+
   it('renders without problems', () => {
-    const wrapper = mount(<Profile profile={fakeProfile} />);
     expect(wrapper.find(Profile)).to.have.length(1);
   });
 
   it('should display the correct username', () => {
-    const wrapper = mount(<Profile profile={fakeProfile} />);
-    expect(wrapper.find('.profile-username').text()).to.equal(fakeProfile.username);
+    expect(wrapper.find('.profile-username').text()).to.equal(fakeStore.getState().profile.username);
   });
 
   it('should display the correct avatar', () => {
-    const wrapper = mount(<Profile profile={fakeProfile} />);
-    expect(wrapper.find('.profile-avatar').html()).to.contain(fakeProfile.avatar);
+    expect(wrapper.find('.profile-avatar').html()).to.contain(fakeStore.getState().profile.avatar);
   });
 });
